requirements.txt
streamlit
numpy
opencv-python
scikit-image
scikit-learn
joblib
Pillow
